Start to house dark-editor state in Dark wasm runtime
