+++ canvases/dark-editor-vue/src/global.d.ts
@@ -1,4 +1,5 @@
 interface Window {
   Darklang: any;
   darklang: any;
+  stateUpdated: any;
 }
