+++ canvases/dark-editor-vue/src/views/MainView.vue
@@ -1,29 +1,14 @@
 <script setup lang="ts">
-import { ref } from 'vue'
-import { onMounted } from '@vue/runtime-core'
 
 import PageHeader from '../components/common/PageHeader.vue'
 import Accordion from '../components/common/Accordion.vue'
-import Conversation from '../components/Conversation.vue'
+//import Conversation from '../components/Conversation.vue'
 
-const systemPromptValue = ref('')
-
-const darklangJSScript: HTMLScriptElement = document.createElement('script')
-darklangJSScript.setAttribute(
-  'src',
-  'http://dark-serve-static.dlio.localhost:11003/editor-bootstrap.js'
-)
-darklangJSScript.setAttribute('defer', '')
-darklangJSScript.addEventListener('load', async () => {
-  const darklang = await window.Darklang.init()
-  window.darklang = darklang
-})
-document.head.appendChild(darklangJSScript)
-
-onMounted(async () => {
-  const response = await fetch('/system-prompt');
-  const responseText = await response.text()
-  systemPromptValue.value = responseText
+let props = defineProps({
+  state: {
+    type: Object, // todo
+    required: true,
+  },
 })
 </script>
 
@@ -42,14 +27,14 @@ onMounted(async () => {
         <textarea
           class="bg-transparent outline-0 p-4 w-full resize-none h-96 max-h-96 overflow-y-auto border-0 text-white"
           ref="prompt"
-          v-model="systemPromptValue"
+          v-model="state.SystemPrompt"
         ></textarea>
       </template>
     </Accordion>
 
     <!-- Conversation -->
-    <div>
+    <!-- <div>
       <Conversation :systemPromptValue="systemPromptValue" />
-    </div>
+    </div> -->
   </main>
 </template>
