+++ canvases/dark-js-wasm-io/README.md
@@ -2,6 +2,6 @@
 
 This demos basic I/O between:
 - the WASM-compiled Darklang backend
-- a wasm.dark that is parsed and then persisted within the running WASM runtime
+- a client.dark that is parsed and then persisted within the running WASM runtime
 - the JS frontend
